moved to booking_page
